tipos de datos de tablas cambiados, nuevas relaciones agregadas

+++ Core.Domain/Entities/Prestamo.cs
@@ -11,13 +11,13 @@ namespace Core.Domain.Entities
     {
         public string NumeroPrestamo { get; set; }
         
-        public int Monto { get; set; }
+        public double Monto { get; set; }
         
-        public int Pago { get; set; }
+        public double Pago { get; set; }
         
         public int Debe { get; set; } //1 , 0
         
-        public int Balance { get; set; }
+        public double Balance { get; set; }
         
         //navigation property
         public int IdProduct { get; set; }
